requirements.txt
pyserial==3.4
docopt==0.6.2
